pyproject.toml
[project]
name = "FEUERnetz"
version = "0.1.5"
description = "Erweiterte Auswertung der Daten aus FEUERnetz"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastexcel>=0.14.0",
    "folium>=0.20.0",
    "marimo>=0.13.6",
    "numpy>=2.2.5",
    "polars-lts-cpu>=1.31.0",
    "pyarrow>=22.0.0",
    "ruff>=0.11.9",
    "seaborn>=0.13.2",
    "utm>=0.8.1",
]
